ASTGen: break the revised stuff out into PDEX
